fix: restore fixer token on reviewer checkpoint persist failure
Restore priorFixerToken when intermediate persistCheckpoint fails after
CreateWorktree clears ownership, and split the oversized ownership
preservation tests into focused prepare-failure, dirty-claim, and
restore-failure files.

Generated-By: looper 0.11.0 (runner=fixer, agent=grok-build)
